Move page query params to center
