+++ src/views/domain-page/config/domain-page-query-params.config.ts
@@ -1,17 +1,18 @@
 import { type PageQueryParam } from '@/hooks/use-page-query-params/use-page-query-params.types';
+import parseDateQueryParam from '@/utils/datetime/parse-date-query-param';
 import { type SortOrder } from '@/utils/sort-by';
 import isWorkflowStatus from '@/views/shared/workflow-status-tag/helpers/is-workflow-status';
 import { type WorkflowStatus } from '@/views/shared/workflow-status-tag/workflow-status-tag.types';
 
-import parseDateQueryParam from '../helpers/parse-date-query-param';
-
-const domainWorkflowsQueryParamsConfig: [
+const domainPageQueryParamsConfig: [
+  // Workflows
   PageQueryParam<'search', string>,
   PageQueryParam<'status', WorkflowStatus | undefined>,
   PageQueryParam<'timeRangeStart', Date | undefined>,
   PageQueryParam<'timeRangeEnd', Date | undefined>,
   PageQueryParam<'sortColumn', string>,
   PageQueryParam<'sortOrder', SortOrder>,
+  // Task Lists
 ] = [
   {
     key: 'search',
@@ -45,4 +46,4 @@ const domainWorkflowsQueryParamsConfig: [
   },
 ] as const;
 
-export default domainWorkflowsQueryParamsConfig;
+export default domainPageQueryParamsConfig;
